decouple to validate_all_urls and validate_changed_urls
