+++ .github/workflows/validate.yml
@@ -24,9 +24,17 @@ jobs:
         python-version: ${{ matrix.python-version }}
         architecture: ${{ matrix.architecture }}
 
-    - name: Install python dependencies
+    - name: flake8 check
       run: |
         python -m pip install -q --upgrade pip
         python -m pip install -q flake8
         flake8
-        python workflow_scripts/validate_urls.py
+
+    - name: Validate changed URLs from this PR
+      run: |
+        python workflow_scripts/validate_changed_urls.py
+
+    - name: Weekly validate all URLs under onnx/tutorials
+      if: (github.event_name == 'schedule') # Only triggered by weekly event
+      run: |
+        python workflow_scripts/validate_all_urls.py
